trying to make a separate camera view
