+++ ToloCam/ComposeViewController.swift
@@ -12,30 +12,27 @@ import ParseUI
 import Bolts
 import AVFoundation
 
-extension UIImage {
-    var uncompressedPNGData: NSData      { return UIImagePNGRepresentation(self)!        }
-    var highestQualityJPEGNSData: NSData { return UIImageJPEGRepresentation(self, 1.0)!  }
-    var highQualityJPEGNSData: NSData    { return UIImageJPEGRepresentation(self, 0.75)! }
-    var mediumQualityJPEGNSData: NSData  { return UIImageJPEGRepresentation(self, 0.5)!  }
-    var lowQualityJPEGNSData: NSData     { return UIImageJPEGRepresentation(self, 0.25)! }
-    var lowestQualityJPEGNSData:NSData   { return UIImageJPEGRepresentation(self, 0.0)!  }
-}
+//extension UIImage {
+ //   var uncompressedPNGData: NSData      { return UIImagePNGRepresentation(self)!        }
+ //   var highestQualityJPEGNSData: NSData { return UIImageJPEGRepresentation(self, 1.0)!  }
+ //   var highQualityJPEGNSData: NSData    { return UIImageJPEGRepresentation(self, 0.75)! }
+ //   var mediumQualityJPEGNSData: NSData  { return UIImageJPEGRepresentation(self, 0.5)!  }
+ //   var lowQualityJPEGNSData: NSData     { return UIImageJPEGRepresentation(self, 0.25)! }
+ //   var lowestQualityJPEGNSData:NSData   { return UIImageJPEGRepresentation(self, 0.0)!  }
+//}
 
-var captureSession : AVCaptureSession?
-var stillImageOutput : AVCaptureStillImageOutput?
-var previewLayer : AVCaptureVideoPreviewLayer?
+//var captureSession : AVCaptureSession?
+//var stillImageOutput : AVCaptureStillImageOutput?
+//var previewLayer : AVCaptureVideoPreviewLayer?
 
 
 class ComposeViewController: UIViewController, UIImagePickerControllerDelegate, UINavigationControllerDelegate, UITextViewDelegate{
 
     @IBOutlet weak var captionTextView: UITextView!
     @IBOutlet weak var previewImage: UIImageView!
-    @IBOutlet weak var cameraView: UIView!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        captionTextView.delegate = self
         
         // Do any additional setup after loading the view.
         
@@ -48,47 +45,6 @@ class ComposeViewController: UIViewController, UIImagePickerControllerDelegate, 
         // Dispose of any resources that can be recreated.
     }
     
-    override func viewDidAppear(animated: Bool) {
-        super.viewDidAppear(animated)
-        previewLayer?.frame = cameraView.bounds
-    }
-    
-    override func viewWillAppear(animated: Bool) {
-        super.viewWillAppear(animated)
-        
-        captureSession = AVCaptureSession()
-        captureSession?.sessionPreset = AVCaptureSessionPreset1920x1080
-        
-        let backCamera = AVCaptureDevice.defaultDeviceWithMediaType(AVMediaTypeVideo)
-        
-        var error : NSError?
-        var input: AVCaptureDeviceInput!
-        do {
-            input = try AVCaptureDeviceInput(device: backCamera)
-        } catch let error1 as NSError {
-            error = error1
-            input = nil
-        }
-        
-        if (error == nil && captureSession?.canAddInput(input) != nil){
-            
-            captureSession?.addInput(input)
-            
-            stillImageOutput = AVCaptureStillImageOutput()
-            stillImageOutput?.outputSettings = [AVVideoCodecKey : AVVideoCodecJPEG]
-            
-            if (captureSession?.canAddOutput(stillImageOutput) != nil){
-                captureSession?.addOutput(stillImageOutput)
-                
-                previewLayer = AVCaptureVideoPreviewLayer(session: captureSession)
-                previewLayer?.videoGravity = AVLayerVideoGravityResizeAspect
-                previewLayer?.connection.videoOrientation = AVCaptureVideoOrientation.Portrait
-                cameraView.layer.addSublayer(previewLayer!)
-                captureSession?.startRunning()
-                
-            }
-            
-            
         }
     
     func addImageTapped(sender: AnyObject) {
@@ -170,5 +126,4 @@ class ComposeViewController: UIViewController, UIImagePickerControllerDelegate, 
     */
     }
 
-}
 }
